% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiec-easi-pkg.R
\name{AGP}
\alias{AGP}
\alias{amgut1.filt}
\alias{amgut2.filt.phy}
\title{American Gut Project}
\format{\enumerate{
 \item amgut1.filt: A matrix with 289 samples (rows) and 127 OTUs (cols).
\item amgut2.filt.phy: A phyloseq object
}}
\source{
http://humanfoodproject.com/americangut/
}
\usage{
data(amgut1.filt)

data(amgut2.filt.phy)
}
\description{
Round 1 and 2 community count datasets from the American Gut Project.
}
