% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_graph.R
\name{as.data.frame.graph}
\alias{as.data.frame.graph}
\title{s3 method for graph to other data types}
\usage{
\method{as.data.frame}{graph}(x, ...)
}
\arguments{
\item{x}{graph adjacency matrix}

\item{...}{Arguments to base as.data.frame}
}
\value{
A data.frame
}
\description{
s3 method for graph to other data types
}
\examples{
# Create a graph and convert to graph adjacency data.frame
g <- make_graph("erdos_renyi", D=5, e=6)
df <- as.data.frame(g)
}
