\name{fitdistr}
\alias{fitdistr}
\title{Fit parameters of a marginal distribution to some data vector}
\usage{
fitdistr(x, densfun, start, ...)
}
\arguments{
  \item{x}{data vector}

  \item{densfun}{string giving distribution function name}

  \item{start}{starting guess for the parameters
  (recommended leaving this out)}

  \item{...}{further arguments to densfun}
}
\description{
Fit parameters of a marginal distribution to some data
vector
}

