% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getOptBeta}
\alias{getOptBeta}
\title{get opt beta}
\usage{
getOptBeta(est)
}
\arguments{
\item{est}{output from \code{spiec.easi} or \code{icov.select}}
}
\value{
Weighted, non-symmetric matrix of model coefficients.
}
\description{
Get the optimal beta matrix when MB neighborhood selection is run with StARS
}

