% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getOptMerge}
\alias{getOptMerge}
\title{get opt merge}
\usage{
getOptMerge(est)
}
\arguments{
\item{est}{output from \code{spiec.easi} or \code{icov.select}}
}
\value{
Weighted, symmetric matrix of edge probability estimates.
}
\description{
Get the optimal merge matrix StARS is run.
}

