% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiec-easi-pkg.R
\docType{data}
\name{hmp2}
\alias{hmp2}
\alias{hmp216S}
\alias{hmp2prot}
\title{Human Microbiome Project 2}
\format{\enumerate{
   \item hmp216S: 16S data, phyloseq object 45 taxa and 47 samples.
   \item hmp2prot: protein data, A phyloseq object, 43 'taxa' and 47 samples.
 }}
\source{
https://www.hmpdacc.org/ihmp/
}
\usage{
data(hmp2)

hmp216S

hmp2prot
}
\description{
Pre-filtered data from the intregrated human microbiome project.
}
\keyword{datasets}
