\name{make_precision_mat}
\alias{make_precision_mat}
\title{Wrapper function for generating a Precision matrix}
\usage{
make_precision_mat(method, D, u, v, ...)
}
\arguments{
  \item{method}{graph generation method}

  \item{D}{graph dimension (number of OTUs)}

  \item{u}{some parameter that controls Prec values}

  \item{v}{some other parameter that controls Prec values}
}
\description{
Wrapper function for generating a Precision matrix
}

