% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_rdiric}
\alias{norm_rdiric}
\title{Normalize via dirichlet sampling}
\usage{
norm_rdiric(x)
}
\arguments{
\item{x}{count data vector}
}
\description{
"Normalize" a count vector by drawing a single sample from a Dirichlet distribution, using the count vector as the prior.
}
\examples{
x <- c(1, 2, 3, 4)
norm_rdiric(x)  # Returns a single sample from Dirichlet distribution
}
