% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_to_total}
\alias{norm_to_total}
\title{Total Sum Normalize}
\usage{
norm_to_total(x)
}
\arguments{
\item{x}{count data vector}
}
\value{
A normalized vector (values sum to 1)
}
\description{
Normalize a count vector by the total sum of that vector
}
\examples{
x <- c(1, 2, 3, 4)
norm_to_total(x)  # Divides each value by sum(x) = 10
}
