% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiec-easi.R
\name{pulsar.params}
\alias{pulsar.params}
\title{pulsar params}
\value{
A list of parameters for pulsar model selection
}
\description{
The values to the \code{pulsar.params}/\code{icov.select.params} argument in the \code{\link{spiec.easi}} function must be a list with values for pulsar model selection parameters.

List of arguments, data type, default. Description
\itemize{
\item thresh, numeric, 0.05. Threshold for StARS criterion.
\item subsample.ratio, numeric, 0.8. Subsample size for StARS.
\item rep.num, numeric, 20. Number of subsamples for StARS.
\item seed, numeric, NULL. Set the random seed for subsample set.
\item ncores, numeric, 1. Number of cores for parallel.
}
With \code{pulsar.select='batch'}, additional arguments:
\itemize{
\item wkdir, dir path, current directory. Working directory for process running jobs.
\item regdir, dir path, temp directory. Directory for storing the registry files.
\item init, string, 'init'. String for differentiating the init registry for batch mode pulsar.
\item conffile, string / file path, ''. Path to config file or string that identifies a default config file.
\item job.res, list, empty list. Named list to specify job resources for an hpc.
\item cleanup, boolean, FALSE. Remove registry files.
}
}
\seealso{
\code{\link{spiec.easi}}
}
