% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{shannon}
\alias{shannon}
\title{compute the shannon entropy from a vector (normalized internally)}
\usage{
shannon(x)
}
\arguments{
\item{x}{data vector}
}
\value{
shannon entropy in base e
}
\description{
Shannon entropy is:
sum [ x_i log(x_i) ]
}
\examples{
x <- c(1, 2, 3, 4)
shannon(x)  # Returns Shannon entropy of normalized vector
}
