% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaRcc.R
\name{sparccboot}
\alias{sparccboot}
\title{SparCC bootstraps}
\usage{
sparccboot(data, sparcc.params = list(), statisticboot, statisticperm, R,
  ncpus = 1, ...)
}
\arguments{
\item{data}{count data matrix}

\item{sparcc.params}{named list of parameters passed to sparcc}

\item{statisticboot}{optional argument to override default bootstrap function}

\item{statisticperm}{optional argument to override default permute function}
}
\value{
sparccboot object. Pass to \code{pval.sparccboot} to get empirical p-values
}
\description{
Bootstrapped Sparcc to get empirical and null distributions for correlations
}

