% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseICov.R
\name{spiec.easi.default}
\alias{spiec.easi.default}
\title{Spiec-Easi pipeline}
\usage{
\method{spiec.easi}{default}(data, method = "glasso",
  sel.criterion = "stars", verbose = TRUE, icov.select = TRUE,
  icov.select.params = list(), ...)
}
\arguments{
\item{data}{non-normalized count OTU/data table with samples on rows and features/OTUs in columns}

\item{method}{estimation method to use as a character string. Currently either 'glasso' or 'mb' (meinshausen-buhlmann)}

\item{sel.criterion}{character string specifying criterion/method for model selection accepts 'stars' [default], 'ric', 'ebic'}

\item{icov.select.params}{list of further arguments to icov.select}

\item{...}{further arguments to sparseiCov}
}
\description{
Spiec-Easi pipeline
}

