% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiec-easi-pkg.R
\docType{package}
\name{SpiecEasi-package}
\alias{SpiecEasi}
\alias{SpiecEasi-package}
\title{SpiecEasi: Sparse Inverse Covariance Estimation for Ecological Association Inference}
\value{
This package provides functions for microbiome network inference and analysis
}
\description{
SpiecEasi provides methods for inferring ecological associations from compositional microbiome data using sparse inverse covariance estimation.
}
\details{
The package implements SPIEC-EASI (SParse InversE Covariance estimation for Ecological Association Inference) methods for microbiome network inference, including glasso and neighborhood selection approaches with stability selection for model selection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zdk123/SpiecEasi}
  \item Report bugs at \url{https://github.com/zdk123/SpiecEasi/issues/}
}

}
\author{
\strong{Maintainer}: Zachary Kurtz \email{zdkurtz@gmail.com}

Authors:
\itemize{
  \item Christian Mueller
  \item Emily Miraldi
  \item Richard Bonneau
}

Other contributors:
\itemize{
  \item Laura Tipton [contributor]
}

}
\keyword{internal}
