% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{xlr}
\alias{xlr}
\title{wrapper for log-ratio transformations with desired pseudo-count
Added for compatability with future SpiecEasi options}
\usage{
xlr(x, fun = clr, pseudofun = function(x) x + 1, ...)
}
\description{
wrapper for log-ratio transformations with desired pseudo-count
Added for compatability with future SpiecEasi options
}
