% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdistr.R
\name{fitdistr}
\alias{fitdistr}
\title{Fit parameters of a marginal distribution to some data vector}
\usage{
fitdistr(x, densfun, start, control, ...)
}
\arguments{
\item{x}{data vector}

\item{densfun}{string giving distribution function name}

\item{start}{starting guess for the parameters (recommended leaving this out)}

\item{control}{control parameters to optim}

\item{...}{further arguments to densfun}
}
\description{
Fit parameters of a marginal distribution to some data vector
}
