% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdistr.R
\name{get_comm_params}
\alias{get_comm_params}
\title{Get the parameters for the OTUs (along mar) of each community}
\usage{
get_comm_params(comm, mar = 2, distr, ...)
}
\arguments{
\item{comm}{community: matrix of counts}

\item{mar}{sample margin (1: "rows", 2: "cols")}

\item{distr}{distribution to fit (see fitdistr)}

\item{...}{arguments passed to fitdistr}
}
\value{
list of parameters
}
\description{
Get the parameters for the OTUs (along mar) of each community
}
