% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_graph.R
\name{graph2prec}
\alias{graph2prec}
\title{Convert a symmetric graph (extension of R matrix class)}
\usage{
graph2prec(Graph, posThetaLims = c(2, 3), negThetaLims = -posThetaLims,
  targetCondition = 100, epsBin = 0.01, numBinSearch = 100)
}
\description{
Has internal rules for converting various graph topologies into the associated
adjancency and, therefore, precision matrix
}

