% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_graph.R
\name{graph2prec2}
\alias{graph2prec2}
\title{DEPRECATED FUNCTION Convert a symmetric graph (extension of R matrix class)}
\usage{
graph2prec2(Graph, theta_min, theta_max, condTheta)
}
\arguments{
\item{u}{some parameter that controls Prec values}

\item{v}{some other parameter that controls Prec values}
}
\description{
Has internal rules for converting various graph topologies into the associated
adjancency and, therefore, precision matrix
}
\keyword{internal}

