% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_graph.R
\name{make_precision_mat}
\alias{make_precision_mat}
\title{DEPREICATED FUNCTION Wrapper function for generating a Precision matrix}
\usage{
make_precision_mat(method, D, u, v, ...)
}
\arguments{
\item{method}{graph generation method}

\item{D}{graph dimension (number of OTUs)}

\item{u}{some parameter that controls Prec values}

\item{v}{some other parameter that controls Prec values}
}
\description{
DEPREICATED FUNCTION Wrapper function for generating a Precision matrix
}
\keyword{internal}

