% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multispieceasi.R
\name{multi.spiec.easi}
\alias{multi.spiec.easi}
\title{Spiec-Easi pipeline for multiple datasets.
All data matrices must be in a list
written by Laura Tipton; 1/5/18}
\usage{
multi.spiec.easi(datalist, method = "glasso", sel.criterion = "stars",
  verbose = FALSE, icov.select = TRUE, icov.select.params = list(), ...)
}
\description{
Spiec-Easi pipeline for multiple datasets.
All data matrices must be in a list
written by Laura Tipton; 1/5/18
}
