% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiec.easi.R
\name{multi.spiec.easi}
\alias{multi.spiec.easi}
\alias{spiec.easi.list}
\title{multi domain SPIEC-EASI}
\usage{
multi.spiec.easi(datalist, method = "glasso", sel.criterion = "stars",
  verbose = TRUE, icov.select = TRUE, icov.select.params = list(), ...)

\method{spiec.easi}{list}(obj, ...)
}
\arguments{
\item{datalist}{list of non-normalized count OTU/data tables (stored in a matrix, data.frame or phyloseq/otu_table) with samples on rows and features/OTUs in columns}

\item{method}{estimation method to use as a character string. Currently either 'glasso' or 'mb' (meinshausen-buhlmann's neighborhood selection)}

\item{sel.criterion}{character string specifying criterion/method for model selection. Accepts 'stars' [default], 'ric', 'ebic' (not recommended for high dimensional data)}

\item{verbose}{flag to show progress messages}

\item{icov.select.params}{list of further arguments to \code{\link{icov.select}}}

\item{...}{further arguments to \code{\link{sparseiCov}} / \code{huge}}
}
\description{
A SPIEC-EASI pipeline for inferring a sparse inverse covariance matrix within and between multiple compositional datasets, under joint sparsity penalty.
}
\details{
Can also run \code{spiec.easi} on a list and S3 will dispatch the proper function.
}
\seealso{
spiec.easi
}
