% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_pseudo}
\alias{norm_pseudo}
\title{Normalize w/ Pseudocount}
\usage{
norm_pseudo(x)
}
\arguments{
\item{x}{count data vector}
}
\description{
add pseudocount before normalizing a count vector
}
\examples{
x <- c(1, 2, 0, 4)
norm_pseudo(x)  # Adds 1 to each value before normalizing
}
