% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_to_total}
\alias{norm_to_total}
\title{Total Sum Normalize}
\usage{
norm_to_total(x)
}
\arguments{
\item{x}{count data vector}
}
\description{
Normalize a count vector by the total sum of that vector
}
