% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaRcc.R
\name{pval.sparccboot}
\alias{pval.sparccboot}
\title{get pvals}
\usage{
pval.sparccboot(x, sided = "both")
}
\arguments{
\item{x}{output from sparccboot}

\item{sided}{compute two sided p-values (one-sided not currently supported)}
}
\description{
Compute empirical p-values from \code{sparccboot} output
}

