% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdistr.R
\name{qqdplot_comm}
\alias{qqdplot_comm}
\title{qq-plot for theoretical vs observed communities}
\usage{
qqdplot_comm(comm, distr, param, plot = TRUE, ...)
}
\description{
qq-plot for theoretical vs observed communities
}

