% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdistr.R
\name{qqdplot_comm}
\alias{qqdplot_comm}
\title{qq-plot for theoretical vs observed communities}
\usage{
qqdplot_comm(comm, distr, param, plot = TRUE, ...)
}
\arguments{
\item{comm}{commutity count matrix}

\item{distr}{character specifying target distribution}

\item{param}{parameter list for fitting the data. Output from \code{get_comm_params}}

\item{plot}{graph the output}

\item{...}{pass arguments to qqplot}
}
\description{
qq-plot for theoretical vs observed communities
}
