% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvdistributions.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Draw samples from multivariate, correlated normal distribution
with counts correlated according to Sigma}
\usage{
rmvnorm(n = 100, mu = rep(0, 10), Sigma = diag(10), tol = 1e-06,
  empirical = TRUE)
}
\arguments{
\item{n}{number of samples to draw}

\item{mu}{mean vector for variables (of length \eqn{D})}

\item{Sigma}{\eqn{DxD} covariance or correlation matrix}

\item{tol}{numerical tolerance for a zero eigenvalue (check for PD of Sigma)}

\item{empirical}{is Sigma the empirical correlation?}
}
\value{
\eqn{Dxn} matrix with Gaussian data
}
\description{
Draw samples from multivariate, correlated normal distribution
with counts correlated according to Sigma
}
