% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaRcc.R
\name{sparcc}
\alias{sparcc}
\title{SparCC}
\usage{
sparcc(data, iter = 20, inner_iter = 10, th = 0.1)
}
\arguments{
\item{data}{count data}

\item{iter}{number of iterations to compute median correlation}

\item{inner_iter}{number of iterations for the inner loop}
}
\description{
reimplementation of SparCC in R
}

