% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseLowRankICov.R
\name{sparseLowRankiCov}
\alias{sparseLowRankiCov}
\title{sparse plus low rank inverse covariance}
\usage{
<<<<<<< HEAD
sparseLowRankiCov(data, npn = FALSE, verbose = FALSE, ...)
=======
sparseLowRankiCov(data, npn = FALSE, verbose = FALSE, cor = FALSE,
  ...)
>>>>>>> update some docs
}
\description{
TODO: add description
}
