% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseLowRankICov.R
\name{sparseLowRankiCov}
\alias{sparseLowRankiCov}
\title{sparse plus low rank inverse covariance}
\usage{
sparseLowRankiCov(data, npn = FALSE, verbose = FALSE, cor = FALSE, ...)
}
\description{
TODO: add description
}
