% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiec-easi.R
\name{spiec.easi}
\alias{spiec.easi}
\alias{spiec.easi.phyloseq}
\alias{spiec.easi.default}
\title{SPIEC-EASI}
\usage{
spiec.easi(data, ...)

\method{spiec.easi}{phyloseq}(data, ...)

\method{spiec.easi}{default}(data, method = "glasso",
  sel.criterion = "stars", verbose = TRUE, pulsar.select = TRUE,
  pulsar.params = list(), icov.select = pulsar.select,
  icov.select.params = pulsar.params, lambda.log = TRUE, ...)
}
\arguments{
\item{data}{OTU count matrix with samples on rows and features/OTUs in columns or a phyloseq object with an OTU count table}

\item{...}{further arguments}

\item{method}{estimation method to use as a character string. Currently either 'glasso' or 'mb' (meinshausen-buhlmann)}

\item{sel.criterion}{character string specifying criterion/method for model selection accepts 'stars' [default] or 'gstars'}

\item{pulsar.select}{flag to use pulsar for model selection}

\item{pulsar.params}{named list of arguments to pass to pulsar. See details.}

\item{icov.select}{defunct. can be set for backwards compatability, but ignored if pulsar.select and/or pulsar.params are given}

\item{icov.select.params}{defunct. can be set for backwards compatability, but ignored if pulsar.select and/or pulsar.params are given}

\item{dist.lambda}{how should lambdas be distributed between lambda.max and lmax*lambda.min.ratio, 'log' or 'linear'.}
}
\description{
Run the whole pipeline
}
