% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseICov.R
\name{spiec.easi}
\alias{spiec.easi}
\title{Spiec-Easi pipeline
Run the whole analysis, from data transformation, iCov estimation and model selection.
Inputs are a non-normalized OTU table and pipeline options.}
\usage{
spiec.easi(obj, ...)
}
\description{
Spiec-Easi pipeline
Run the whole analysis, from data transformation, iCov estimation and model selection.
Inputs are a non-normalized OTU table and pipeline options.
}

