% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiec-easi-pkg.R
\docType{package}
\name{SpiecEasi-package}
\alias{SpiecEasi}
\alias{SpiecEasi-package}
\title{SpiecEasi: Sparse Inverse Covariance for Ecological Statistical Inference}
\description{
Estimate networks from the precision matrix of compositional microbial abundance data.
}
\author{
\strong{Maintainer}: Zachary Kurtz \email{zdkurtz@gmail.com}

Authors:
\itemize{
  \item Christian Mueller
  \item Emily Miraldi
  \item Richard Bonneau
}

Other contributors:
\itemize{
  \item Laura Tipton [contributor]
}

}
\keyword{internal}
