% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{stars.roc}
\alias{stars.roc}
\alias{stars.pr}
\title{stars.roc, stars.pr}
\usage{
stars.roc(optmerge, theta, verbose = TRUE, plot = TRUE, ll = 15)

stars.pr(optmerge, theta, verbose = TRUE, plot = TRUE, ll = 15)
}
\arguments{
\item{optmerge}{the optimal 'merge' matrix selected by stars}

\item{theta}{the true graph or precision matrix}

\item{verbose}{display messages}

\item{plot}{graph the output}

\item{ll}{number of points for the plot}
}
\description{
Plot a ROC (reciever operator characteristic) or a Precision-Recall curve
along the stars 'confidence path'. Each edge is a number in [0,1],
which is on the fraction of inferred graphs over subsamples in which that edge
appeared in stars.
}
