% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{symBeta}
\alias{symBeta}
\title{sym beta}
\usage{
symBeta(beta, mode = "ave")
}
\arguments{
\item{beta}{square coefficient matrix}

\item{mode}{how to symmetrize, see details}
}
\value{
a symmetric coefficient matrix
}
\description{
Symmetrize a beta (coefficient) matrix, ie. selected from MB neighborhood selection
}
\details{
Mode can be:
\itemize{
 \item ave:  Arithmetic average between the two possible values of beta
 \item maxabs: The maximum [absolute] value between the two values
 \item upper:  Take the values from the upper triangle
 \item lower:  Take the values from the lower triangle
}
}
