% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{triu}
\alias{triu}
\alias{tril}
\alias{triu2diag}
\title{Functions for triangular matrices}
\usage{
triu(x, k = 1)

tril(x, k = 1)

triu2diag(x, diagval = 0)
}
\description{
Get or symmeterize the upper/lower triangle of a symmetric matrix with the other side zeroed out
}
